requirements.txt
numpy==1.23.3
scipy==1.9.1
pyaml